Fix all compiler and CMake warnings for clean build
Compiler warnings fixed:
- -Wswitch in flann.hpp: added default case to switch
- -Wsign-compare in nn_index.h: changed int to size_t
- -Wformat in kmeans_opencl_index.h: cast radius to double
- -Wterminate in nn_opencl_index.h: removed throw from destructor
- -Wunused-parameter in allocator.h and kmeans_search_cooperative.cuh
- -Wsign-compare in kdtree_cuda_3d_index.cu: added static_cast
- -Wc++17-extensions in kmeans_cuda_index.h: replaced structured bindings

CMake warnings fixed:
- cmake_minimum_required: updated to VERSION 3.5...3.28
- CMP0135: added DOWNLOAD_EXTRACT_TIMESTAMP TRUE
- CMP0148: replaced FindPythonInterp with find_package(Python)
- CMP0026: replaced LOCATION property with generator expression

+++ src/cpp/flann/algorithms/kdtree_cuda_3d_index.cu
@@ -572,7 +572,7 @@ int KDTreeCuda3dIndex<Distance >::radiusSearchGpu(const Matrix<ElementType>& que
 
     int buffer_index=0;
     for( size_t i=0; i<queries.rows; i++ ) {
-        for( size_t j=0; j<counts_host[i]; j++ ) {
+        for( size_t j=0; j<static_cast<size_t>(counts_host[i]); j++ ) {
             dists[i][j]=dists_temp[buffer_index];
             indices[i][j]=indices_temp[buffer_index];
             ++buffer_index;
